[연경] 등산 후기 게시판 완료
MountainReviewPhoto : 수정 폼에서 기존(이미 업로드 된) 이미지 불러와서 추가할 수 있도록 기능 추가
모든 id Type int -> Long으로 변경

+++ src/main/java/himedia/hpm_spring/repository/vo/ReviewPhotoVo.java
@@ -17,7 +17,7 @@ import lombok.ToString;
 @ToString
 public class ReviewPhotoVo {
 
-    private int id;					//	reviewPhoto 고유번호
+    private Long id;					//	reviewPhoto 고유번호
     
     @JsonProperty("file_name")
     private String fileName;		//	파일 이름
@@ -29,5 +29,5 @@ public class ReviewPhotoVo {
     private Date updateDate;		//	갱신 시간
     
     @JsonProperty("reviews_id")
-    private int reviewsId;			//	review ID
+    private Long reviewsId;			//	review ID
 }
